Extract KomaArtist classes to separate file koma_artist.py
